requirements.txt
streamlit
google-generativeai
requests
python-dotenv
pandas
openpyxl
gspread
oauth2client
supabase
